advanced_preprocessing_system/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
albumentations>=1.3.0
scikit-image>=0.18.0
matplotlib>=3.5.0
ultralytics>=8.0.0
segment-anything>=1.0
transformers>=4.21.0